Add Subject Alternative Names extension
