+++ pki_tools/types/certificate/__init__.py
@@ -1,116 +1,20 @@
-import binascii
-from collections import defaultdict
 from datetime import datetime
-from typing import List, Union, Optional
+from typing import Union, Optional
 
 
 from cryptography import x509
-from cryptography.hazmat._oid import NameOID
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import padding, ec, dsa, rsa
 from cryptography.hazmat.primitives.asymmetric.types import (
     CertificatePublicKeyTypes,
 )
 
-from pydantic import BaseModel, Field, ConfigDict
+from pydantic import BaseModel, ConfigDict
 
-import pki_tools
+from pki_tools.types.certificate.name import Name
 from pki_tools.types.certificate.extensions import Extensions
-from pki_tools.types import _is_pem_str, PemCert, _byte_to_hex
+from pki_tools.types import _byte_to_hex
 
-
-class Subject(BaseModel):
-    """
-    Subject type describes certificate subject or issuer.
-    The attributes are following the
-    [RFC5280#Section-4.1.2.4](https://datatracker.ietf.org/doc/html/rfc5280#section-4.1.2.4)
-
-    Note that every attribute is a list of string in order to support
-    multivalued RDNs.
-
-    Attributes:
-        c -- Country Name (2.5.4.6)
-        o -- Organization Name (2.5.4.10)
-        ou -- Organizational Unit Name (2.5.4.11)
-        dnq -- Distinguished Name Qualifier (2.5.4.46)
-        s -- State Or Province Name (2.5.4.8)
-        cn -- Common Name (2.5.4.3)
-        serial -- Serial Number (2.5.4.5)
-        ln -- Locality Name (2.5.4.7)
-        t -- Title (2.5.4.12)
-        sn -- Surname (2.5.4.4)
-        gn -- Given Name (2.5.4.42)
-        i -- Initials (2.5.4.43)
-        p -- Pseudonym (2.5.4.65)
-        gq -- Generation Qualifier (2.5.4.44)
-        dc -- Domain Component (0.9.2342.19200300.100.1.25)
-    """
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    c: List[str] = Field(alias=NameOID.COUNTRY_NAME.dotted_string, default=[])
-    o: List[str] = Field(
-        alias=NameOID.ORGANIZATION_NAME.dotted_string, default=[]
-    )
-    ou: List[str] = Field(
-        alias=NameOID.ORGANIZATIONAL_UNIT_NAME.dotted_string, default=[]
-    )
-    dnq: List[str] = Field(
-        alias=NameOID.DN_QUALIFIER.dotted_string, default=[]
-    )
-    s: List[str] = Field(
-        alias=NameOID.STATE_OR_PROVINCE_NAME.dotted_string, default=[]
-    )
-    cn: List[str] = Field(alias=NameOID.COMMON_NAME.dotted_string, default=[])
-    serial: List[str] = Field(
-        alias=NameOID.SERIAL_NUMBER.dotted_string, default=[]
-    )
-
-    ln: List[str] = Field(
-        alias=NameOID.LOCALITY_NAME.dotted_string, default=[]
-    )
-    t: List[str] = Field(alias=NameOID.TITLE.dotted_string, default=[])
-    sn: List[str] = Field(alias=NameOID.SURNAME.dotted_string, default=[])
-    gn: List[str] = Field(alias=NameOID.GIVEN_NAME.dotted_string, default=[])
-    i: List[str] = Field(alias=NameOID.INITIALS.dotted_string, default=[])
-    p: List[str] = Field(alias=NameOID.PSEUDONYM.dotted_string, default=[])
-    gq: List[str] = Field(
-        alias=NameOID.GENERATION_QUALIFIER.dotted_string, default=[]
-    )
-    dc: List[str] = Field(
-        alias=NameOID.DOMAIN_COMPONENT.dotted_string, default=[]
-    )
-
-    @classmethod
-    def from_cryptography(cls, name: x509.Name):
-        subject = defaultdict(set)
-        for attribute in name:
-            for att in name.get_attributes_for_oid(attribute.oid):
-                subject[att.oid.dotted_string].add(att.value)
-        return cls(**subject)
-
-    def to_crypto_name(self) -> x509.Name:
-        name_list = []
-        for attr_name in vars(self):
-            vals = getattr(self, attr_name)
-            if not vals:
-                continue
-
-            oid = Subject.model_fields[attr_name].alias
-            for val in vals:
-                name_list.append(
-                    x509.NameAttribute(x509.ObjectIdentifier(oid), val)
-                )
-
-        return x509.Name(name_list)
-
-    def __str__(self):
-        attrs = []
-        for a in self.model_dump():
-            for val in getattr(self, a):
-                attrs.append(f"{a.upper()} = {val}")
-        subject_str = ", ".join(attrs)
-        return f"Subject: {subject_str}"
 
 
 class SignatureAlgorithm(BaseModel):
@@ -184,9 +88,9 @@ class TbsCertificate(BaseModel):
     version: int
     serial_number: int
     signature_algorithm: SignatureAlgorithm
-    issuer: Subject
+    issuer: Name
     validity: Validity
-    subject: Subject
+    subject: Name
     subject_public_key_info: SubjectPublicKeyInfo
     extensions: Optional[Extensions]
 
@@ -226,10 +130,7 @@ class Certificate(TbsCertificate):
     signature_value: str
 
     @classmethod
-    def parse_certificate(cls, cert: [x509.Certificate, PemCert]):
-        if _is_pem_str(cert):
-            cert = pki_tools.cert_from_pem(cert)
-
+    def parse_certificate(cls, cert: x509.Certificate):
         return cls(
             version=cert.version.value,
             serial_number=cert.serial_number,
@@ -237,12 +138,12 @@ class Certificate(TbsCertificate):
                 algorithm=cert.signature_hash_algorithm,
                 parameters=cert.signature_algorithm_parameters,
             ),
-            issuer=Subject.from_cryptography(cert.issuer),
+            issuer=Name.from_cryptography(cert.issuer),
             validity=Validity(
                 not_before=cert.not_valid_before,
                 not_after=cert.not_valid_after,
             ),
-            subject=Subject.from_cryptography(cert.subject),
+            subject=Name.from_cryptography(cert.subject),
             subject_public_key_info=SubjectPublicKeyInfo.from_cryptography(
                 cert.public_key()
             ),
